chore: add "how to get started" instructions (#9)
I was confused where to look and how to get the discussify sidebar up (I didn't see the white floating action button at all, so I didn't think clicking on the extension was doing anything). To help others, I added a line that would have helped me figure out how to get this working. Even better if we could add an explanatory image or gif to this page, and maybe a tooltip or animation on the discuss button when it first appears to call attention to it.